Reorganize snakebids directory
- Move code from __init__.py to bids.py
- Use `mkinit` tool to manage __init__.py file imports
- Move `snakemake_io.py` and `outputs.py` to `utils/` directory

Address issue #62
