+++ snakebids/tests/__init__.py
@@ -1,0 +1,5 @@
+__submodules__ = []
+# <AUTOGEN_INIT>
+__all__ = []
+
+# </AUTOGEN_INIT>
